Move `atomic_wrapper` and `thread_suspension` into `detail` folder.
