+++ include/concore2full/detail/bulk_spawn_frame_base.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "concore2full/c/atomic_wrapper.h"
 #include "concore2full/c/spawn.h"
 #include "concore2full/c/task.h"
-#include "concore2full/c/thread_suspension.h"
+#include "concore2full/detail/atomic_wrapper.h"
 #include "concore2full/detail/core_types.h"
+#include "concore2full/detail/thread_suspension.h"
 #include "concore2full/this_thread.h"
 
 #include <memory>
